+++ demo/pages/ui/swap.tsx
@@ -59,6 +59,7 @@ const SwapWidgetPage: NextPage = () => {
     symbol: 'ETH',
     logoURI: 'https://assets.relay.link/icons/currencies/eth.png'
   })
+
   const { setWalletFilter } = useWalletFilter()
   const { setShowAuthFlow, primaryWallet } = useDynamicContext()
   const { theme } = useTheme()
@@ -220,16 +221,16 @@ const SwapWidgetPage: NextPage = () => {
             lockChainId={singleChainMode ? 8453 : undefined}
             singleChainMode={singleChainMode}
             supportedWalletVMs={supportedWalletVMs}
-            sponsoredTokens={
-              toToken?.chainId !== 1
-                ? [
-                    '8453:0x833589fcd6edb6e08f4c7c32d4f71b54bda02913',
-                    '42161:0xaf88d065e77c8cc2239327c5edb3a432268e5831',
-                    '10:0x0b2c639c533813f4aa9d7837caf62653d097ff85',
-                    '1:0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48'
-                  ]
-                : undefined
-            }
+            sponsoredTokens={[
+              '792703809:EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v',
+              '130:0x078d782b760474a361dda0af3839290b0ef57ad6',
+              '103665049:0xdba34672e30cb065b1f93e3ab55318768fd6fef66c15942c9f7cb846e2f900e7::usdc::USDC',
+              '8453:0x833589fcd6edb6e08f4c7c32d4f71b54bda02913',
+              '43114:0xb97ef9ef8734c71904d8002f8b6bc66dd9c48a6e',
+              '137:0x3c499c542cef5e3811e1192ce70d8cc03d5c3359',
+              '42161:0xaf88d065e77c8cc2239327c5edb3a432268e5831',
+              '10:0x0b2c639c533813f4aa9d7837caf62653d097ff85'
+            ]}
             // popularChainIds={[]}
             // disableInputAutoFocus={true}
             toToken={toToken}